refactor(s3): move s3 utility to config directory
The s3 utility function was relocated from infrastructure/utils to config directory to better reflect its purpose as a configuration-related utility. This improves code organization and maintainability.
